// frontend/src/app/model/rodeInfo.ts
export interface RoadnetWorkInventory {
  idpci?: any;
  Authority?: any;
  NetworkID?: any;
  branchName?: any;
  sectionID?: any;
  ROW?: any;
  functionalClass?: any;
  PavementType?: any;
  SegLength?: any;
  PavementWidth?: any;
  Shoulderwidth?: any;
  noOfLanes?: any;
  Medianwidth?: any;
  divided?: any;
  Con_main_Reh_Year?: any;
  Age?: any;
  trafficVolume?: any;
  from?: any;
  X?: any;
  to?: any;
  X1?: any;
  Y1?: any;
  submittedBy?: any;
  submittedOn?: any;
}